split the map_commands crate into contract, wire and runtime so transport types are kept out of execution signatures
