--- host/crates/map_commands/src/lib.rs
@@ -1,6 +1,0 @@
-pub mod dispatch;
-pub mod domain;
-pub mod wire;
-
-#[cfg(test)]
-mod tests;
