--- host/crates/map_commands/src/tests/mod.rs
@@ -1,2 +1,0 @@
-mod dispatch_tests;
-mod wire_serde_tests;
