+++ host/crates/map_commands_contract/src/command_descriptor.rs
