+++ host/crates/map_commands_contract/src/holon_command.rs
