+++ host/crates/map_commands_contract/src/space_command.rs
