+++ host/crates/map_commands_runtime/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "map_commands_runtime"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+map_commands_contract = { workspace = true }
+holons_core = { workspace = true }
+base_types = { workspace = true }
+core_types = { workspace = true }
+
+[dev-dependencies]
+tokio = { version = "1", features = ["rt", "macros"] }
+serde_json = { workspace = true }
