+++ host/crates/map_commands_runtime/src/holon_handler.rs
@@ -2,7 +2,7 @@ use base_types::{BaseValue, MapString};
 use core_types::HolonError;
 use holons_core::reference_layer::{HolonReference, ReadableHolon, WritableHolon};
 
-use crate::domain::{HolonAction, HolonCommand, MapResult, ReadableHolonAction, WritableHolonAction};
+use map_commands_contract::{HolonAction, HolonCommand, MapResult, ReadableHolonAction, WritableHolonAction};
 
 /// Handles holon-scoped commands.
 pub async fn handle_holon(command: HolonCommand) -> Result<MapResult, HolonError> {
