+++ host/crates/map_commands_runtime/src/lib.rs
@@ -6,3 +6,6 @@ mod transaction_handler;
 
 pub use runtime::Runtime;
 pub use runtime_session::RuntimeSession;
+
+#[cfg(test)]
+mod tests;
