+++ host/crates/map_commands_wire/src/ipc_envelope.rs
